extract AutocompleteSuggestionItem into a separate component
